% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{candsetsize}
\alias{candsetsize}
\title{candsetsize}
\usage{
candsetsize(npos, zMI)
}
\arguments{
\item{npos}{Regression coefficient associated with the eigenvectors}

\item{zMI}{Standardized Moran coefficient of regression residuals}
}
\value{
An integer indicating the ideal size of the candidate set
}
\description{
Determine the ideal size of the candidate set
}
\references{
Chun, Yongwan, Daniel A. Griffith, Monghyeon Lee, Parmanand
Sinha (2016): Eigenvector selection with stepwise regression techniques
to construct eigenvector spatial filters. Journal of Geographical
Systems, 18, pp. 67 – 85.
}
\seealso{
\code{\link{lmfilter}}, \code{\link{getEVs}}
}
\author{
Sebastian Juhl
}
