% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emp.pfunc}
\alias{emp.pfunc}
\title{Calculate empirical \emph{p}-Value for Standardized
Moran Coefficient}
\usage{
emp.pfunc(draws, z, alternative)
}
\arguments{
\item{draws}{Number of simulation iterations}

\item{z}{Observed value for the standardized Moran coefficient}

\item{alternative}{Optional: alternative hypothesis to be assessed. Default
is 'two-sided'}
}
\value{
Returns the empirical \emph{p}-value associated with the supplied
(standardized) Moran coefficient based on a simulated null distribution
}
\description{
Derive empirical \emph{p}-value for the standardized
Moran coefficient
}
\seealso{
\code{\link{lmfilter}}, \code{\link{pfunc}}
}
\author{
Sebastian Juhl
}
