% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakedataset.R
\docType{data}
\name{fakedata}
\alias{fakedata}
\alias{fakedataset}
\alias{W}
\title{Synthetic Dataset}
\format{
An object of class \code{data.frame} with 100 rows and 8 columns.

An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 100 columns.
}
\usage{
data(fakedata)

W
}
\value{
The file contains two objects:
\tabular{lcl}{
\code{fakedataset} \tab \tab a synthetic dataset\cr
\code{W} \tab \tab an artificial spatial connectivity matrix\cr
}
}
\description{
An artificially generated cross-sectional dataset together with
an accompanying binary connectivity matrix \emph{\strong{W}}. The
\eqn{n=100} units are located on a regular grid and \emph{\strong{W}}
is defined according to rook's adjacency definition of contiguity.
The synthetic data can be used to illustrate the functionality
of this package.
}
\examples{

data(fakedata)
head(fakedataset)
dim(W)
}
\keyword{dataset}
\keyword{datasets}
