% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakedataset.R
\docType{data}
\name{fakedataset}
\alias{fakedataset}
\title{Fake Dataset}
\format{
An object of class \code{data.frame} with 100 rows and 8 columns.
}
\usage{
data(toydata)
}
\value{
The file contains two objects:
\tabular{lcl}{
\code{fakedataset}\tab \tab a synthetic dataset of dimensionality \eqn{100 x 8}\cr
\code{W}\tab\tab a binary connectivity matrix
}
}
\description{
An artificially generated cross-sectional dataset together with
an accompanying binary connectivity matrix \emph{\strong{W}} to
illustrate the use of the package. The \eqn{10 x 10} matrix connects
the \eqn{n=100} units based on rook's adjacency definition of contiguity.
}
\examples{

data(fakedata)
head(fakedataset)
dim(W)
}
\keyword{dataset}
