% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfilter.R
\name{lmfilter}
\alias{lmfilter}
\title{Eigenvector-Based Spatial Filtering in Linear Regression Models}
\usage{
lmfilter(
  y,
  x = NULL,
  W,
  objfn = "MI",
  MX = FALSE,
  sig = 0.05,
  bonferroni = TRUE,
  positive = TRUE,
  ideal.setsize = FALSE,
  alpha = 0.25,
  tol = 0.1,
  na.rm = TRUE
)
}
\arguments{
\item{y}{Vector of regressands}

\item{x}{Matrix of regressors (default = NULL)}

\item{W}{Spatial connectivity matrix}

\item{objfn}{Specifies the objective function to be used for eigenvector
selection. Possible criteria are: the maximization of the
adjusted R-squared ('R2'), minimization of residual autocorrelation ('MI'),
significance level of candidate eigenvectors ('p'), or
all eigenvectors in the candidate set ('all')}

\item{MX}{Covariates used to construct the projection matrix (TRUE/ FALSE)}

\item{sig}{Significance level to be used for eigenvector selection
if \code{objfn='p'}}

\item{bonferroni}{Bonferroni adjustment for the significance level
(TRUE/ FALSE)}

\item{positive}{Restrict search to eigenvectors associated with positive
levels of spatial autocorrelation (TRUE/ FALSE)}

\item{ideal.setsize}{If \code{positive=TRUE}, uses the formula proposed in
Chun et al. (2016) to determine the ideal size of the candidate set
(TRUE/ FALSE)}

\item{tol}{If \code{objfn='MI'}, determines the amount of remaining residual
autocorrelation at which the eigenvector selection terminates}

\item{na.rm}{Removes missing values in covariates (TRUE/ FALSE)}
}
\value{
An object of class \code{spfilter} containing the following
information:
\describe{
\item{Estimates}{Summary statistics of the parameter estimates}
\item{varcovar}{Estimated variance-covariance matrix}
\item{selvecs}{Matrix of selected eigenvectors}
\item{evMI}{Moran coefficient of eigenvectors in the spatial filter}
\item{moran}{Residual autocorrelation for the initial and the filtered model}
\item{fit}{Adjusted R-squared of the initial and the filtered model}
\item{residuals}{Model residuals}
\item{other}{A list providing supplementary information:
\describe{
\item{ncandidates}{Number of candidate eigenvectors considered}
\item{nev}{Number of selected eigenvectors}
\item{sel_id}{ID of selected eigenvectors}
\item{sf}{Vector representing the spatial filter}
\item{sfMI}{Moran coefficient of the spatial filter}
\item{model}{Class of the regression model}
\item{MX}{TRUE/ FALSE: covariates included in projection matrix
\emph{\strong{M}}}
\item{dependence}{Filtered for positive or negative spatial dependence}
\item{objfn}{Selection criteria specified in the objective function of
the stepwise regression procedure}
\item{bonferroni}{TRUE/ FALSE: bonferroni-adjusted significance level
(if \code{objfn='p'})}
\item{siglevel}{If \code{objfn='p'}: actual (unadjusted/ adjusted)
significance level}
}
}
}
}
\description{
Performs spatial filtering with eigenvectors in a
linear regression framework
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
\references{
Tiefelsdorf, Michael and Daniel A. Griffith (2007):
Semiparametric filtering of spatial autocorrelation: the eigenvector
approach. Environment and Planning A: Economy and Space, 39 (5):
pp. 1193 - 1221.

Chun, Yongwan, Daniel A. Griffith, Monghyeon Lee, Parmanand
Sinha (2016): Eigenvector selection with stepwise regression techniques
to construct eigenvector spatial filters. Journal of Geographical
Systems, 18, pp. 67 – 85.

Le Gallo, Julie and Antonio Páez (2013): Using synthetic
variables in instrumental variable estimation of spatial series models.
Environment and Planning A: Economy and Space, 45 (9): pp. 2227 - 2242.

Tiefelsdorf, Michael and Barry Boots (1995): The Exact Distribution
of Moran's I. Environment and Planning A: Economy and Space, 27 (6):
pp. 985 - 999.
}
