% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.sf.R
\name{MI.sf}
\alias{MI.sf}
\title{Moran Coefficient of the Spatial Filter}
\usage{
MI.sf(gamma, evMI)
}
\arguments{
\item{gamma}{regression coefficient associated with the eigenvectors}

\item{evMI}{Moran coefficient of selected eigenvectors}
}
\value{
Moran coefficient of the spatial filter.
}
\description{
Computes the Moran coefficient for the spatial filter.
}
\references{
Le Gallo, Julie and Antonio Páez (2013): Using synthetic
variables in instrumental variable estimation of spatial series models.
Environment and Planning A: Economy and Space, 45 (9): pp. 2227 - 2242.
}
\seealso{
\code{\link{lmFilter}}, \code{\link{getEVs}}, \code{\link{MI.ev}}
}
\author{
Sebastian Juhl
}
