% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.vec.R
\name{MI.vec}
\alias{MI.vec}
\title{Moran Test for Spatial Autocorrelation in a Vector}
\usage{
MI.vec(x, W, alternative = "greater")
}
\arguments{
\item{x}{a vector}

\item{W}{spatial connectivity matrix}

\item{alternative}{specification of alternative hypothesis as 'greater' (default),
'lower', or 'two.sided'}
}
\value{
Returns an object of class \code{data.frame} that contains the
following information:
\tabular{lcl}{
\code{I}\tab \tab observed value of the Moran coefficient\cr
\code{EI}\tab\tab expected value of Moran's I\cr
\code{VarI}\tab \tab variance of Moran's I (under normality)\cr
\code{zI}\tab\tab standardized Moran coefficient\cr
\code{pI}\tab\tab \emph{p}-value of the test statistic
}
}
\description{
Tests for the presence of spatial autocorrelation
in a vector as indicated by the Moran coefficient. The variance
is calculated under the normality assumtion.
}
\references{
Cliff, Andrew D. and John K. Ord (1981): Spatial Processes:
Models & Applications. Pion, London.

Bivand, Roger S. and David W. S. Wong (2018): Comparing
Implementations of Global and Local Indicators of Spatial Association.
TEST 27: pp. 716 - 748.

Tiefelsdorf, Michael and Barry Boots (1995): The Exact Distribution
of Moran's I. Environment and Planning A: Economy and Space, 27 (6):
pp. 985 - 999.
}
\seealso{
\code{\link{getMoran}}
}
\author{
Sebastian Juhl
}
