% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partialR2}
\alias{partialR2}
\title{Partial R-squared}
\usage{
partialR2(y, x, evecs)
}
\arguments{
\item{y}{Vector of regressands}

\item{x}{Matrix of regressors}

\item{evecs}{Selected eigenvectors for inclusion in the spatial filter}
}
\value{
Vector of partial R-squared values for each eigenvector
}
\description{
Partial R-squared of selected eigenvectors
}
\seealso{
\code{\link{lmfilter}}, \code{\link{getEVs}}
}
\author{
Sebastian Juhl
}
