% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pfunc}
\alias{pfunc}
\title{Calculate \emph{p}-Value for Standardized Moran Coefficient}
\usage{
pfunc(z, alternative)
}
\arguments{
\item{z}{Standardized Moran coefficient}

\item{alternative}{Optional: alternative hypothesis to be assessed. Default
is 'two-sided'}
}
\value{
Returns the \emph{p}-value associated with the supplied
(standardized) Moran coefficient
}
\description{
Derive \emph{p}-value for the standardized Moran coefficient
}
\seealso{
\code{\link{lmfilter}}, \code{\link{emp.pfunc}}
}
\author{
Sebastian Juhl
}
