% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{star}
\alias{star}
\title{Graphical Illustration of Significance Levels}
\usage{
star(p)
}
\arguments{
\item{p}{\emph{p}-value}
}
\value{
Returns the empirical \emph{p}-value associated with the supplied
(standardized) Moran coefficient based on a simulated null distribution
}
\description{
Illustrate significance levels for output
}
\seealso{
\code{\link{lmfilter}}, \code{\link{pfunc}}, \code{\link{emp.pfunc}}
}
\author{
Sebastian Juhl
}
