% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.ev.R
\name{vif.ev}
\alias{vif.ev}
\title{Variance Inflation Factor for Eigenvectors}
\usage{
vif.ev(x = NULL, evecs)
}
\arguments{
\item{x}{vector/ matrix of regressors (default = NULL)}

\item{evecs}{(selected) eigenvectors}
}
\value{
Returns a vector containing the VIF for each eigenvector.
}
\description{
Calculate the variance inflation factor (VIF) for
the eigenvectors in the spatial filter.
}
\examples{
data("fakedata")
evecs <- getEVs(W=W)$vectors
evecs <- evecs[,1:10]
vif.ev(x=fakedataset$x1,evecs=evecs)

}
\seealso{
\code{\link{lmFilter}}, \code{\link{getEVs}}
}
\author{
Sebastian Juhl
}
