% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEVs.R
\name{getEVs}
\alias{getEVs}
\title{Eigenfunction Decomposition of a (Transformed) Spatial Connectivity
Matrix}
\usage{
getEVs(W, covars = NULL)
}
\arguments{
\item{W}{spatial connectivity matrix}

\item{covars}{vector/ matrix of regressors included in the construction
of the projection matrix \emph{\strong{M}}}
}
\value{
A list containing the following objects:
\describe{
\item{\code{vectors}}{A matrix containing \emph{n} eigenvectors}
\item{\code{values}}{A vector of the corresponding eigenvalues}
}
}
\description{
Extract eigenvectors and corresponding eigenvalues from
the matrix \emph{\strong{MWM}}, where \emph{\strong{M}} denotes a symmetric
and idempotent projection matrix and \emph{\strong{W}} is the spatial
connectivity matrix.
}
\details{
The eigenfunctions obtained by \code{getEVs}
can be used to perform supervised eigenvector selection and to
manually create a spatial filter.

If \emph{\strong{W}} is not symmetric, \code{getEVs} symmetrizes the
matrix by: \eqn{0.5 * (W + t(W))}.
}
\seealso{
\code{\link{lmFilter}}, \code{\link{MI.ev}}, \code{\link{MI.sf}},
\code{\link{vif.ev}}, \code{\link{partialR2}}
}
\author{
Sebastian Juhl
}
