% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEVs.R
\name{getEVs}
\alias{getEVs}
\title{Extract Eigenvectors and Eigenvalues}
\usage{
getEVs(W, covars = NULL)
}
\arguments{
\item{W}{Spatial connectivity matrix}

\item{covars}{Design matrix of regressors included in the construction
of the projection matrix \emph{\strong{M}}}
}
\value{
A list containing the following objects:
\describe{
\item{\code{vectors}}{A matrix containing \emph{n} eigenvectors}
\item{\code{values}}{A vector of the corresponding eigenvalues}
}
}
\description{
Extract eigenvectors and the corresponding eigenvalues from
the matrix \emph{\strong{MWM}}, where \emph{\strong{M}} denotes a symmetric
and idempotent projection matrix and \emph{\strong{W}} is the spatial
connectivity matrix.
}
\seealso{
\code{\link{lmfilter}}, \code{\link{MI.ev}}, \code{\link{MI.sf}},
\code{\link{vif.ev}}, \code{\link{partialR2}}
}
\author{
Sebastian Juhl
}
