% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMoran.R
\name{getMoran}
\alias{getMoran}
\title{Moran Coefficient for Residual Spatial Autocorrelation}
\usage{
getMoran(
  y,
  x = NULL,
  fitted.values = NULL,
  W,
  alternative = "greater",
  boot = NULL
)
}
\arguments{
\item{y}{vector of regressands}

\item{x}{matrix of regressors (default = NULL)}

\item{fitted.values}{a vector of fitted values from a regression model (default = NULL)}

\item{W}{spatial connectivity matrix}

\item{alternative}{specification of alternative hypothesis as 'greater' (default),
'lower', or 'two.sided'}

\item{boot}{optional integer specifying the number of bootstrap iterations
to compute the variance. If NULL (default), variance calculated under normality}
}
\value{
A \code{data.frame} object with the following elements:
\tabular{lcl}{
\code{I}\tab \tab observed value of the Moran coefficient\cr
\code{EI}\tab\tab expected value of Moran's I\cr
\code{VarI}\tab \tab variance of Moran's I\cr
\code{zI}\tab\tab standardized Moran coefficient\cr
\code{pI}\tab\tab \emph{p}-value of the test statistic
}
}
\description{
This function assesses the degree of spatial
autocorrelation in regression residuals by means of the Moran
coefficient.
}
\references{
Tiefelsdorf, Michael and Barry Boots (1995): The Exact Distribution
of Moran's I. Environment and Planning A: Economy and Space, 27 (6):
pp. 985 - 999.

Tiefelsdorf, Michael and Daniel A. Griffith (2007):
Semiparametric filtering of spatial autocorrelation: the eigenvector
approach. Environment and Planning A: Economy and Space, 39 (5):
pp. 1193 - 1221.
}
\seealso{
\code{\link{lmFilter}}, \code{\link{MI.vec}}
}
\author{
Sebastian Juhl
}
