% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMoran.R
\name{getMoran}
\alias{getMoran}
\title{Moran Test for Residual Spatial Autocorrelation}
\usage{
getMoran(
  y,
  x = NULL,
  fitted.values = NULL,
  W,
  alternative = "greater",
  boot = NULL
)
}
\arguments{
\item{y}{vector of regressands}

\item{x}{vector/ matrix of regressors (default = NULL)}

\item{fitted.values}{a vector of fitted values from a regression model (default = NULL)}

\item{W}{spatial connectivity matrix}

\item{alternative}{specification of alternative hypothesis as 'greater' (default),
'lower', or 'two.sided'}

\item{boot}{optional integer specifying the number of bootstrap iterations
to compute the variance. If NULL (default), variance calculated under normality}
}
\value{
A \code{data.frame} object with the following elements:
\tabular{lcl}{
\code{I}\tab \tab observed value of the Moran coefficient\cr
\code{EI}\tab\tab expected value of Moran's I\cr
\code{VarI}\tab \tab variance of Moran's I\cr
\code{zI}\tab\tab standardized Moran coefficient\cr
\code{pI}\tab\tab \emph{p}-value of the test statistic
}
}
\description{
This function assesses the degree of spatial
autocorrelation present in regression residuals by means of the Moran
coefficient.
}
\details{
The function directly uses fitted values to compute the residuals
whenever they are supplied. If neither \code{x} nor \code{fitted.values}
are supplied, the function assumes a linear intercept-only model and
calculates model residuals accordingly.
}
\note{
Calculations are based on Cliff and Ord (1981) and Upton and Fingleton
(1985). See also Tiefelsdorf (2000) and Griffith et al. (2019).
}
\examples{
data(fakedata)
y <- fakedataset$x1
x <- fakedataset$x3

Moran <- getMoran(y=y,x=x,W=W,alternative='greater')
Moran

}
\references{
Tiefelsdorf, Michael (2000): Modelling Spatial Processes.
The Identification and Analysis of Spatial Relationships in Regression
Residuals by Means of Moran's I. Springer, Berlin.

Griffith, Daniel A., Yongwan Chun, Bin Li (2019): Spatial Regression
Analysis Using Eigenvector Spatial Filtering. Elsevier Academic Press,
London.
}
\seealso{
\code{\link{lmFilter}}, \code{\link{MI.vec}}
}
\author{
Sebastian Juhl
}
