% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMoran.R
\name{getMoran}
\alias{getMoran}
\title{Moran Test for Residual Spatial Autocorrelation}
\usage{
getMoran(resid, x = NULL, W, alternative = "greater", boot = NULL)
}
\arguments{
\item{resid}{residual vector}

\item{x}{vector/ matrix of regressors (default=NULL)}

\item{W}{spatial connectivity matrix}

\item{alternative}{specification of alternative hypothesis as 'greater' (default),
'lower', or 'two.sided'}

\item{boot}{optional integer specifying the number of simulation iterations to
compute the variance. If NULL (default), variance calculated under assumed normality}
}
\value{
A \code{data.frame} object with the following elements:
\describe{
\item{\code{I}}{observed value of the Moran coefficient}
\item{\code{EI}}{expected value of Moran's I}
\item{\code{VarI}}{variance of Moran's I}
\item{\code{zI}}{standardized Moran coefficient}
\item{\code{pI}}{\emph{p}-value of the test statistic}
}
}
\description{
This function assesses the degree of spatial
autocorrelation present in regression residuals by means of the Moran
coefficient.
}
\details{
The function assumes an intercept-only model if \code{x=NULL}.
Furthermore, \code{getMoran} automatically symmetrizes the matrix
\emph{\strong{W}} by: 1/2 * (\emph{\strong{W}} + \emph{\strong{W}}').
}
\note{
Calculations are based on Cliff and Ord (1981) and Upton and Fingleton
(1985). See also Tiefelsdorf (2000) and Griffith et al. (2019).
}
\examples{
data(fakedata)
y <- fakedataset$x1
x <- fakedataset$x2

resid <- y - x \%*\% solve(crossprod(x)) \%*\% crossprod(x,y)

(Moran <- getMoran(resid=resid,x=x,W=W,alternative="greater"))

}
\references{
Tiefelsdorf, Michael (2000): Modelling Spatial Processes.
The Identification and Analysis of Spatial Relationships in Regression
Residuals by Means of Moran's I. Springer, Berlin.

Griffith, Daniel A., Yongwan Chun, Bin Li (2019): Spatial Regression
Analysis Using Eigenvector Spatial Filtering. Elsevier Academic Press,
London.

Cliff, Andrew D. and John K. Ord (1981): Spatial Processes:
Models & Applications. Pion, London.

Upton, Graham J. G. and Bernard Fingleton (1985): Spatial Data Analysis
by Example, Volume 1.New York, Wiley.
}
\seealso{
\code{\link{lmFilter}}, \code{\link{glmFilter}}, \code{\link{MI.vec}}
}
\author{
Sebastian Juhl
}
