% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFilter.R, R/methods.R
\name{lmFilter}
\alias{lmFilter}
\alias{summary.spfilter}
\alias{plot.spfilter}
\title{Unsupervised Spatial Filtering with Eigenvectors in Linear Regression Models}
\usage{
lmFilter(
  y,
  x = NULL,
  W,
  objfn = "MI",
  MX = FALSE,
  sig = 0.05,
  bonferroni = TRUE,
  positive = TRUE,
  ideal.setsize = FALSE,
  alpha = 0.25,
  tol = 0.1,
  na.rm = TRUE
)

\method{summary}{spfilter}(obj, EV = FALSE)

\method{plot}{spfilter}(obj)
}
\arguments{
\item{y}{vector of regressands}

\item{x}{vector/ matrix of regressors (default=NULL)}

\item{W}{spatial connectivity matrix}

\item{objfn}{specifies the objective function to be used for eigenvector
selection. Possible criteria are: the maximization of the
adjusted R-squared ('R2'), minimization of residual autocorrelation ('MI'),
significance level of candidate eigenvectors ('p'), or
all eigenvectors in the candidate set ('all')}

\item{MX}{covariates used to construct the projection matrix (TRUE/ FALSE)}

\item{sig}{significance level to be used for eigenvector selection
if \code{objfn='p'}}

\item{bonferroni}{Bonferroni adjustment for the significance level
(TRUE/ FALSE)}

\item{positive}{restrict search to eigenvectors associated with positive
levels of spatial autocorrelation (TRUE/ FALSE)}

\item{ideal.setsize}{if \code{positive=TRUE}, uses the formula proposed in
Chun et al. (2016) to determine the ideal size of the candidate set
(TRUE/ FALSE)}

\item{alpha}{a value in (0,1] indicating the range of candidate eigenvectors
according to their associated level of spatial autocorrelation, see e.g.,
Griffith (2003)}

\item{tol}{if \code{objfn='MI'}, determines the amount of remaining residual
autocorrelation at which the eigenvector selection terminates}

\item{na.rm}{remove missing values in covariates (TRUE/ FALSE)}

\item{obj}{an object of class \code{spfilter}}

\item{EV}{display summary statistics for selected eigenvectors (TRUE/ FALSE)}
}
\value{
An object of class \code{spfilter} containing the following
information:
\tabular{lcl}{
\code{Estimates}\tab \tab summary statistics of the parameter estimates\cr
\code{varcovar}\tab\tab estimated variance-covariance matrix\cr
\code{EV}\tab\tab a matrix with summary statistics of selected eigenvectors\cr
\code{selvecs}\tab\tab matrix of selected eigenvectors\cr
\code{evMI}\tab\tab Moran coefficient of all eigenvectors\cr
\code{moran}\tab\tab residual autocorrelation for the initial and the
filtered model\cr
\code{fit}\tab\tab adjusted R-squared of the initial and the filtered model\cr
\code{residuals}\tab\tab initial and filtered model residuals\cr
\code{other}\tab\tab a list providing supplementary information:\cr
\tab\tab \describe{
\item{\code{ncandidates}}{number of candidate eigenvectors considered}
\item{\code{nev}}{number of selected eigenvectors}
\item{\code{sel_id}}{ID of selected eigenvectors}
\item{\code{sf}}{vector representing the spatial filter}
\item{\code{sfMI}}{Moran coefficient of the spatial filter}
\item{\code{model}}{class of the regression model}
\item{\code{MX}}{TRUE/ FALSE: include covariates in the projection matrix
\emph{\strong{M}}}
\item{\code{dependence}}{filtered for positive or negative spatial dependence}
\item{\code{objfn}}{selection criteria specified in the objective function of
the stepwise regression procedure}
\item{\code{bonferroni}}{TRUE/ FALSE: Bonferroni-adjusted significance level
(if \code{objfn='p'})}
\item{\code{siglevel}}{if \code{objfn='p'}: actual (unadjusted/ adjusted)
significance level}
}
}
}
\description{
This function implements the eigenvector-based semiparametric
spatial filtering approach in a linear regression framework using OLS.
Eigenvectors are selected by an unsupervised steppwise regression
technique. Supported selection criteria are the minimization of residual
autocorrelation, maximization of model fit, and the statistical significance
of eigenvectors. Alternatively, all eigenvectors in the candidate set
can be included as well.
}
\details{
If \emph{\strong{W}} is not symmetric, it gets symmetrized by
0.5 * (\emph{\strong{W}} + \emph{\strong{W}}') the eigenfunction decomposition.

If \code{MX=TRUE}, the function uses the covariates specified in the argument
\code{x} to construct the following projection matrix:

\emph{\strong{M} = \strong{I} - \strong{X} (\strong{X}'\strong{X})^-1\strong{X}'}

Eigenvectors from \emph{\strong{MWM}} using this specification of
\emph{\strong{M}} are not only mutually uncorrelated but also orthogonal
to the included regressors. Alternatively, if \code{MX=FALSE}, the projection
matrix becomes \emph{\strong{M} = \strong{I} - \strong{1}
(\strong{1}'\strong{1})^-1\strong{1}'}, where \emph{\strong{1}} is a vector of
ones. Griffith and Tiefelsdorf (2007) show how the choice of the appropriate
\emph{\strong{M}} depends on the underlying process that generates the spatial
dependence.
}
\examples{
data(fakedata)
y <- fakedataset$x1
X <- cbind(fakedataset$x2,fakedataset$x3,fakedataset$x4)

res <- lmFilter(y=y,x=X,W=W,objfn='MI',positive=FALSE)
print(res)
summary(res,EV=TRUE)

E <- res$selvecs
(ols <- coef(lm(y~X+E)))
coef(res)

}
\references{
Tiefelsdorf, Michael and Daniel A. Griffith (2007):
Semiparametric filtering of spatial autocorrelation: the eigenvector
approach. Environment and Planning A: Economy and Space, 39 (5):
pp. 1193 - 1221.

Griffith, Daniel A. (2003): Spatial Autocorrelation and Spatial Filtering:
Gaining Understanding Through Theory and Scientific Visualization.
Berlin/ Heidelberg, Springer.

Chun, Yongwan, Daniel A. Griffith, Monghyeon Lee, Parmanand
Sinha (2016): Eigenvector selection with stepwise regression techniques
to construct eigenvector spatial filters. Journal of Geographical
Systems, 18, pp. 67 – 85.

Le Gallo, Julie and Antonio Páez (2013): Using synthetic
variables in instrumental variable estimation of spatial series models.
Environment and Planning A: Economy and Space, 45 (9): pp. 2227 - 2242.

Tiefelsdorf, Michael and Barry Boots (1995): The Exact Distribution
of Moran's I. Environment and Planning A: Economy and Space, 27 (6):
pp. 985 - 999.
}
\seealso{
\code{\link{glmFilter}}, \code{\link{getEVs}}, \code{\link{getMoran}}
}
