% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MI.ev}
\alias{MI.ev}
\title{MI.ev}
\usage{
MI.ev(W, evals)
}
\arguments{
\item{W}{Spatial connectivity matrix}

\item{evals}{Eigenvalues of selected eigenvectors}
}
\value{
Returns a vector containing the Moran coefficients of each
selected eigenvector
}
\description{
Moran coefficient for each eigenvector in the spatial filter
}
\references{
Le Gallo, Julie and Antonio Páez (2013): Using synthetic
variables in instrumental variable estimation of spatial series models.
Environment and Planning A, 45 (9): pp. 2227 - 2242.

Tiefelsdorf, Michael and Barry Boots (1995): The Exact Distribution
of Moran's I. Environment and Planning A: Economy and Space, 27 (6):
pp. 985 - 999.
}
\seealso{
\code{\link{lmfilter}}, \code{\link{getEVs}}, \code{\link{MI.sf}}
}
\author{
Sebastian Juhl
}
