% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vif.ev}
\alias{vif.ev}
\title{Variance Inflation Factor for Eigenvectors}
\usage{
vif.ev(x = NULL, evecs)
}
\arguments{
\item{x}{Matrix of regressors (default = NULL)}

\item{evecs}{Selected eigenvectors for inclusion in the spatial filter}
}
\value{
Returns a vector containing the VIF for each selected eigenvector
}
\description{
Calculate the variance inflation factor (VIF) for
the eigenvectors in the spatial filter
}
\seealso{
\code{\link{lmfilter}}
}
\author{
Sebastian Juhl
}
